fixup(m2): dedupe catalog error helper; export ApiError; document admin-store catalog endpoint choice
- stores/mcpShared.ts: CatalogErrorEnvelope + toCatalogError extracted
- api/index.ts: ApiError now part of public surface for instanceof checks
- mcpStore.fetchCatalog: JSDoc explaining workspace-scoped catalog endpoint
