+++ frontend/packages/core/src/stores/index.ts
@@ -14,6 +14,7 @@ export { useWorkspaceStore, type WorkspaceStore } from './workspaceStore'
 export { useAttachmentStore, type UploadingFile } from './attachmentStore'
 export { useMcpStore, type McpStore } from './mcpStore'
 export { useWorkspaceMcpStore, type WorkspaceMcpStore } from './workspaceMcpStore'
+export { type CatalogErrorEnvelope, toCatalogError } from './mcpShared'
 export { useProvidersStore } from './providersStore'
 export { useModelsStore } from './modelsStore'
 export { useOrgModelSettingsStore } from './orgModelSettingsStore'
